+++ options.go
@@ -21,11 +21,19 @@ var (
 //	func init() {
 //	    testo.Options(myplugin.OutputDir("..."))
 //	}
-func Options(options ...testoplugin.Option) {
+//
+// It returns an empty struct to enable the following usage:
+//
+//	var _ = testo.Options(...)
+//
+// This is similar to [For] and slighlty more concise than using init.
+func Options(options ...testoplugin.Option) struct{} {
 	globalOptionsMutex.Lock()
 	defer globalOptionsMutex.Unlock()
 
 	globalOptions = append(globalOptions, options...)
+
+	return struct{}{}
 }
 
 func getOptions() []testoplugin.Option {