+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- `testo.Options` now returns an empty struct to enable `var _ = testo.Options(...)` usage.
+
 ## [1.2.0] - 2026-05-21
 
 ### Added